+++ protected/modules/admin/controllers/SlidesController.php
@@ -67,7 +67,8 @@ class SlidesController extends Controller
 		$model = Slides::model()->findByPk($id);
 		$name = $model->attributes['image'];
 		$this->loadModel($id)->delete();
-		unlink(Yii::app()->basePath.'/../upload/images/'.$name);
+		if(file_exists(Yii::app()->basePath.'/../upload/images/'.$name) && $name->image != '')
+			unlink(Yii::app()->basePath.'/../upload/images/'.$name);
 		Yii::app()->user->setFlash('success', translate('Xóa thành công.'));
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 		$this->redirect(PIUrl::createUrl('/admin/slides'));
